refactor: `MobileTestController`를 컴포넌트로 분리
